1111
예상버그 없 음
편의성 충분히 추가

이제 남은 시간 동안 야외 이벤트 같은 것 만 열심히 추가하자

+++ lehoo/Assets/Script/UI/UI_skill_info.cs
@@ -141,7 +141,7 @@ public class GameJsonData
 
   public int QuestType = 0;
   public int Cult_Phase = 0;
-  public int Cult_Progress = 0;
+  public float Cult_Progress = 0;
   public int Cult_SabbatSector = 0;
   public Vector2 Cult_RitualTile = new Vector2();
   public int Cult_CoolTime = 0;
